fix(core): use `cwd` for locating workspace root
Closes #3484
Closes #3909
